% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sampler}
\alias{sampler}
\title{Sample n numbers from (1, ..., 10) with replacement}
\usage{
sampler(n)
}
\arguments{
\item{n}{The number of samples to be drawn.}
}
\value{
A vector of length \code{n}.
}
\description{
Draw a given number of samples from the integers from 1 to 10 inclusive, with
replacement.
}
\examples{
n <- 6
sampler(n)

}

