% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sumDivide}
\alias{sumDivide}
\title{Divide the sum of two numbers by another number}
\usage{
sumDivide(x, y, z)
}
\arguments{
\item{x}{The first of the two numbers to be added.}

\item{y}{The second of the two numbers to be added.}

\item{z}{The number by which the sum of \code{x} and \code{y} should be
divided.}
}
\value{
The sum of the first two arguments, divided by the third argument.
}
\description{
Add together two numbers and divide the answer by a third number.
}
\examples{
a <- 12
b <- 24
c <- 4
sumDivide(a, b, c) # Evaluates to 9

}

